requirements.txt
crewai>=0.28.0
requests>=2.31.0
pytest>=8.0.0
python-dotenv>=1.0.0
langchain-ollama>=0.2.0